fix: guard iFruit.Update and fix config list float cast
- iFruitMMI.OnTick: _iFruit.Update() now inside its own try/catch.
  An exception there escaped OnTick and SHVDN aborted the whole
  script, freezing the phone UI / stuck money notifications.
- MenuConfig float list: Math.Round returns double, stored as object,
  then unboxed as (float) -> InvalidCastException on every scroll of
  the multiplier lists. Store floats so the cast matches.

+++ iFruit/MenuConfig.cs
@@ -130,7 +130,12 @@ namespace MMI_SP.iFruit
 
             for (float i = startValue; i < stopValue; i += increment)
             {
-                values.Add(Math.Round(i, 1, MidpointRounding.AwayFromZero));
+                // Math.Round returns double. Storing boxed doubles then
+                // unboxing as (float) in the OnListChange handler below
+                // throws InvalidCastException ("Specified cast is not valid")
+                // on every scroll of this list. Store floats instead so the
+                // (float) cast matches the boxed type.
+                values.Add((float)Math.Round(i, 1, MidpointRounding.AwayFromZero));
                 if (!found)
                     if (Math.Round(value, 1, MidpointRounding.AwayFromZero) == Math.Round(i, 1, MidpointRounding.AwayFromZero))
                         found = true;
